+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - `X-Client-Name` and `X-Client-Version` headers
+- Add SEP0010 transaction challenge builder.
+    ```
+    client = Stellar::Client.default_testnet
+    client.build_challenge_tx(server: server, client: user, anchor_name: anchor, timeout: timeout) 
+    ```
 
 ## [0.7.0] - 2019-04-26
 ### Added